Add licence declarations with reuse
This PR makes the project license declare licenses according to
https://reuse.software/ so that it passes the reuse lint.

Even though it is many affected files, the pr is scroll friendly and contains no logic.
It should be a quick general overview before merging it.

Basically, add adds license headers in spdx-format, as per the resuse-specification.
It also adds a missing reuse LICENSES file, ofl 1.1, and the .reuse/dep5 which takes care of the licening for binary files etc.

Note: the reuse badge in the README will be green when the project is public
Note: some files created with the react-app was a bit unsure,
see the comment in dep5, and it will be ok for now.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>